refactor(core): Extract UserGroupExternalManager for user external-id operations
Move user external-id CRUD and enable/disable from UserGroupManager into
UserGroupExternalManager; AccessControlManager delegates to the new class.
Always set externalId on addUser without a null check.
